[FE 1.0] Extract NativeObjCNameOverridesChecker from NativeObjCNameChecker
